+++ contrib/storage-jdbc/src/main/java/org/apache/drill/exec/store/jdbc/JdbcStorageConfig.java
@@ -37,7 +37,6 @@ import org.apache.drill.exec.store.security.UsernamePasswordCredentials;
 public class JdbcStorageConfig extends AbstractSecuredStoragePluginConfig {
 
   public static final String NAME = "jdbc";
-  public static final String JDBC_CLICKHOUSE_PREFIX = "jdbc:clickhouse:";
 
   private final String driver;
   private final String url;
@@ -94,11 +93,6 @@ public class JdbcStorageConfig extends AbstractSecuredStoragePluginConfig {
   @JsonIgnore
   public UsernamePasswordCredentials getUsernamePasswordCredentials() {
     return new UsernamePasswordCredentials(credentialsProvider);
-  }
-
-  @JsonIgnore
-  public boolean isClickhouse() {
-    return url.startsWith(JDBC_CLICKHOUSE_PREFIX);
   }
 
   @Override
